fix(ResourceHeader): rename to F0ResourceHeader to clear the stable DoD name-prefix check
The naming requirement added by #5137 checks the component's folder against
the "F0" + uppercase convention. ResourceHeader was the last stable-tagged
component still missing that prefix (it was cleared of every other DoD gap by
#4989), so it regressed back onto the debt list. Rename the folder, exported
symbol, and type to F0ResourceHeader, and update every internal consumer.
