+++ packages/react/src/components/dialog-alike/F0Drawer/__stories__/F0Drawer.mdx
@@ -101,7 +101,7 @@ Drawers can include module information and tabs in the header for complex conten
 
 ### With Resource Header
 
-You can embed a `ResourceHeader` component within the drawer content for a consistent header experience.
+You can embed a `F0ResourceHeader` component within the drawer content for a consistent header experience.
 
 <Canvas of={F0DrawerStories.WithResourceHeader} />
