+++ packages/react/src/experimental/F0MeetingCard/__stories__/F0MeetingCard.mdx
@@ -57,7 +57,7 @@ it with `attendeesDisplay` when a surface genuinely needs the other one.
 
 Only `maxAvatars` faces are shown (3 by default); the rest collapse into a `+N`
 counter. **Hover that counter and the collapsed attendees are listed** with their
-name and email — the same behaviour as the avatar list in `ResourceHeader`, since
+name and email — the same behaviour as the avatar list in `F0ResourceHeader`, since
 both go through `F0AvatarList`.
 
 <Canvas of={Stories.InProgress} />
