+++ packages/react/src/patterns/F0Dialog/__stories__/F0Modal.stories.tsx
@@ -6,8 +6,8 @@ import { expect, within } from "storybook/test"
 import { F0Button } from "@/components/F0Button"
 import { ApplicationFrame } from "@/patterns/ApplicationFrame"
 import ApplicationFrameStoryMeta from "@/patterns/ApplicationFrame/index.stories"
-import { ResourceHeader } from "@/patterns/ResourceHeader"
-import { Default as ResourceHeaderDefault } from "@/patterns/ResourceHeader/index.stories"
+import { F0ResourceHeader } from "@/patterns/F0ResourceHeader"
+import { Default as ResourceHeaderDefault } from "@/patterns/F0ResourceHeader/index.stories"
 import {
   OnePersonListItem,
   OnePersonListItemProps,
@@ -414,9 +414,9 @@ export const WithResourceHeader: Story = {
       href: "/timeoff",
     },
     children: (
-      <ResourceHeader
+      <F0ResourceHeader
         {...(ResourceHeaderDefault.args as ComponentProps<
-          typeof ResourceHeader
+          typeof F0ResourceHeader
         >)}
         primaryAction={undefined}
         secondaryActions={undefined}
