+++ packages/react/src/patterns/F0ResourceHeader/index.mdx
@@ -339,7 +339,7 @@ When hovering over a metadata field, one or more of these actions may be display
 ## Import
 
 ```tsx
-import { ResourceHeader } from "@factorialco/f0-react"
+import { F0ResourceHeader } from "@factorialco/f0-react"
 ```
 
 ## Examples
