+++ packages/react/src/ui/ButtonGroup/variants.ts
@@ -25,7 +25,7 @@ export const buttonGroupVariants = cva({
      * Orientation + responsive stacking, encoded as ONE axis:
      * - `none`: always a horizontal row.
      * - `sm` / `md`: stack as a column below the named *viewport* breakpoint and
-     *   become a row at/above it (matches Card footer `sm`, ResourceHeader `md`).
+     *   become a row at/above it (matches Card footer `sm`, F0ResourceHeader `md`).
      * - `container-md`: stack below the *container* `@md` breakpoint (28rem / 448px);
      *   requires an ancestor with `@container` (matches F0CardHorizontal).
      */
